v12-production-api/pyproject.toml
[project]
name = "studybuddy-v12"
version = "12.0.0"
description = "StudyBuddy v12 - Production API with Authentication & Monitoring"
requires-python = ">=3.12,<3.14"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",  # File uploads
    "aiofiles>=24.1.0",  # Async file operations
    "websockets>=13.0",  # WebSocket support

    # LangChain ecosystem (1.2+ for create_agent API)
    "langchain>=1.2.0",
    "langchain-core>=1.2.0",
    "langchain-openai>=1.1.0",
    "langchain-community>=0.4.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-qdrant>=1.0.0",

    # LangGraph for agents
    "langgraph>=1.0.0",

    # MCP integration (v11)
    "langchain-mcp-adapters>=0.2.0",  # LangChain <-> MCP bridge
    "mcp>=1.0.0",  # Model Context Protocol SDK

    # Vector database
    "qdrant-client>=1.12.0",

    # Utilities
    "python-dotenv>=1.0.0",

    # Persistence
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",

    # Advanced Retrieval (from v9)
    "langchain-experimental>=0.3.0",  # SemanticChunker
    "cohere>=5.0.0",  # Reranking
    "rank-bm25>=0.2.2",  # BM25 sparse retrieval

    # Document processing
    "pypdf>=4.0.0",  # PDF parsing
    "unstructured>=0.15.0,<0.18.31",  # Markdown and other formats (0.18.31+ adds numba which fails on Python 3.12)

    # Web content extraction (v11 - for Fetch connector post-processing)
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",

    # Authentication (v12)
    "pyjwt[crypto]>=2.8.0",  # JWT tokens
    "bcrypt>=4.0.0",  # Password hashing

    # Rate limiting (v12)
    "slowapi>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]
